backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
supabase==2.11.0
google-genai==1.5.0
langgraph==0.2.60
pydantic==2.10.3
python-jose[cryptography]==3.3.0
httpx==0.28.1
python-multipart==0.0.18
